Moving sorting functionality to separate module under hnsw
