+++ iris-mpc-cpu/src/hnsw/graph/mod.rs
@@ -1,4 +1,3 @@
-pub mod batcher_sort;
 pub mod graph_store;
 pub mod layered_graph;
 pub mod neighborhood;
